+++ examples/fulldemo/plugins/local_demo/__init__.py
@@ -21,7 +21,7 @@ __version__ = "0.1.0"
 bp = Blueprint(__plugin_name__, __plugin_name__)
 
 
-@bp.route('/')
+@bp.route("/")
 def index():
     return jsonify(dict(hello="localdemo"))
 
@@ -30,28 +30,33 @@ def br():
     local = LocalStorage()
     local.set(
         "nowtime",
-        time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(time.time()))
+        time.strftime("%Y-%m-%d %H:%M:%S", time.localtime(time.time())),
     )
 
 
 def view_limit(path):
     if request.method == "GET":
-        return jsonify(dict(method='GET', status=200, path=path))
+        return jsonify(dict(method="GET", status=200, path=path))
     else:
         return jsonify(dict(method=request.method, status=403, path=path))
 
 
 def page_not_found(error):
-    return jsonify(dict(status=404, msg="Not Found Page")),404
+    return jsonify(dict(status=404, msg="Not Found Page")), 404
 
 
 def register():
     return {
-        "tep": dict(code=u'<p>hello local-demo(from html code)</p>', html='localdemo/title.html'),
-        'hep': dict(before_request=br),
-        'bep': dict(blueprint=bp, prefix='/localdemo'),
-        'vep': dict(rule='/limit/<path>', view_func=view_limit, methods=['GET', 'POST']),
-        'filter': [('demo_filter2', lambda x: 'test-filter')],
-        'errhandler': {404: page_not_found},
-        'tcp': dict(timestamp=int(time.time())),
+        "tep": dict(
+            code=u"<p>hello local-demo(from html code)</p>",
+            html="localdemo/title.html",
+        ),
+        "hep": dict(before_request=br),
+        "bep": dict(blueprint=bp, prefix="/localdemo"),
+        "vep": dict(
+            rule="/limit/<path>", view_func=view_limit, methods=["GET", "POST"]
+        ),
+        "filter": [("demo_filter2", lambda x: "test-filter")],
+        "errhandler": {404: page_not_found},
+        "tcp": dict(timestamp=int(time.time())),
     }