+++ tests/test_installer.py
@@ -7,26 +7,33 @@ from flask_pluginkit.utils import check_url
 
 
 class PITest(unittest.TestCase):
-
     def setUp(self):
-        self.pi = PluginInstaller('.')
+        self.pi = PluginInstaller(".")
 
     def test_addPlugin(self):
         urls = [
-            'https://static.saintic.com/download/thirdApp/JQueryAvatarPreviewCut.zip',
-            'https://static.saintic.com/download/thirdApp/syncthing-linux-amd64-v0.14.45.tar.gz',
-            'https://codeload.github.com/saintic/jwt/zip/master',
-            'https://codeload.github.com/staugur/Flask-PluginKit/tar.gz/v3.0.0',
+            "https://static.saintic.com/download/thirdApp/JQueryAvatarPreviewCut.zip",
+            "https://static.saintic.com/download/thirdApp/syncthing-linux-amd64-v0.14.45.tar.gz",
+            "https://codeload.github.com/saintic/jwt/zip/master",
+            "https://codeload.github.com/staugur/Flask-PluginKit/tar.gz/v3.0.0",
         ]
         pkg = os.path.join(self.pi.plugin_abspath, "jwt-master")
         # local
-        res = self.pi.addPlugin(method='local', filepath='/tmp/1.tar', remove=True)
+        res = self.pi.addPlugin(
+            method="local", filepath="/tmp/1.tar", remove=True
+        )
         self.assertEqual(res["code"], 1)
-        res = self.pi.addPlugin(method='local', filepath='/tmp/1.tbz', remove=True)
+        res = self.pi.addPlugin(
+            method="local", filepath="/tmp/1.tbz", remove=True
+        )
         self.assertEqual(res["code"], 1)
-        res = self.pi.addPlugin(method='local', filepath='/tmp/1.tgz', remove=True)
+        res = self.pi.addPlugin(
+            method="local", filepath="/tmp/1.tgz", remove=True
+        )
         self.assertEqual(res["code"], 1)
-        res = self.pi.addPlugin(method='local', filepath='/tmp/1.zip', remove=True)
+        res = self.pi.addPlugin(
+            method="local", filepath="/tmp/1.zip", remove=True
+        )
         self.assertEqual(res["code"], 1)
         # remote
         res = self.pi.addPlugin(url=urls[2])
@@ -52,5 +59,6 @@ class PITest(unittest.TestCase):
         self.assertEqual(res["code"], 0)
         self.assertFalse(os.path.isdir(pkg))
 
-if __name__ == '__main__' and not os.getenv("TRAVIS"):
+
+if __name__ == "__main__" and not os.getenv("TRAVIS"):
     unittest.main()